requirements.txt
discord.py~=2.3.2
pytz~=2024.1
pymongo~=4.7.0

requests~=2.32.3
msgpack~=1.0.2
nonebot~=1.8.4
Pillow~=9.0.1
pycryptodome~=3.14.1
zhconv~=1.4.3

pyinstaller~=6.6.0